座標単位をEMUからptに統一
全座標値（位置・サイズ・フォントサイズ・線幅・マージン等）をEMUからpt（1pt=12700EMU）に変換。
LLMが直感的にレイアウトやフォントサイズを把握できるようにする。

+++ internal/pptx/callout.go
@@ -45,8 +45,8 @@ func resolveCalloutPointer(geom *xmlPrstGeom, pos *Position) *Point {
 		}
 	}
 
-	px := pos.X + pos.W/2 + a1*pos.W/100000
-	py := pos.Y + pos.H/2 + a2*pos.H/100000
+	px := pos.X + pos.W/2 + float64(a1)*pos.W/100000
+	py := pos.Y + pos.H/2 + float64(a2)*pos.H/100000
 
 	return &Point{X: px, Y: py}
 }
